requirements.txt
requests
beautifulsoup4
streamlit
google-api-python-client
google-auth